Fix Ambience foliage color
NOTE:
- Azalea, Cherry and Pale Oak are not affected
- For some reason now Mangroves use the default tint in Mojang source
