+++ src/main/resources/meteor-client.mixins.json
@@ -11,6 +11,7 @@
     "AbstractClientPlayerEntityMixin",
     "AbstractFurnaceScreenHandlerMixin",
     "AbstractFurnaceScreenMixin",
+    "AbstractSignBlockEntityRendererMixin",
     "AbstractSignEditScreenAccessor",
     "AbstractSignEditScreenMixin",
     "ArmorFeatureRendererMixin",
@@ -22,6 +23,7 @@
     "BeaconScreenMixin",
     "BiomeColorsMixin",
     "BlockCollisionSpliteratorMixin",
+    "BlockColorsMixin",
     "BlockEntityRenderDispatcherMixin",
     "BlockEntityTypeAccessor",
     "BlockHitResultAccessor",
@@ -95,7 +97,6 @@
     "FireworksSparkParticleSubMixin",
     "FishingBobberEntityMixin",
     "FluidRendererMixin",
-    "FoliageColorsMixin",
     "GameOptionsMixin",
     "GameRendererMixin",
     "HandledScreenAccessor",
@@ -161,7 +162,6 @@
     "ServerResourcePackLoaderMixin",
     "ShapeIndexBufferAccessor",
     "ShulkerBoxBlockMixin",
-    "AbstractSignBlockEntityRendererMixin",
     "SimpleEntityLookupAccessor",
     "SimpleOptionMixin",
     "SlimeBlockMixin",
